Attempting GPS
Using UART in order to transmit GGA style message to parse.
